Changed the calculation of the Weighted Projection
Earlier we calculated the weight as the change in Market Capitalization,
and calculated the product of price-change and ratio-in-total-market-cap.
Instead of that, we had to calculate the ratio in total shares, since market_cap = price * num_of_shares, and we should not use the price twice in the calculation.

After this change, the square error dropped by 50%.